requirements.txt
kivy>=2.3.0
kivymd>=1.2.0
google-generativeai
python-dotenv
Pillow
